http: Reintegrate roundtrip into outbound
